lib/formFields: move form fields to lib for reuse
