+++ src/components/accounts/add/importjson.js
@@ -1,13 +1,12 @@
 import React from 'react';
-import Dropzone from 'react-dropzone';
 import { connect } from 'react-redux'
 import { Field, reduxForm, reset } from 'redux-form'
-import TextField from 'material-ui/TextField'
 import { Card, CardActions, CardHeader, CardText } from 'material-ui/Card'
 import FlatButton from 'material-ui/FlatButton'
 import FontIcon from 'material-ui/FontIcon'
 
 import { cardSpace } from 'lib/styles'
+import { renderFileField, renderTextField } from 'lib/formFields'
 import { Row, Col } from 'react-flexbox-grid/lib/index'
 
 import Immutable from 'immutable'
@@ -16,32 +15,6 @@ import { importWallet } from 'store/accountActions'
 import { required } from 'lib/validators'
 import log from 'loglevel'
 import AccountShow from '../show'
-
-const renderFileField = ({ input, name, meta: { touched, error } }) => { 
-
-    const files = input.value
-    const onDrop = ( filesToUpload, e ) => input.onChange(filesToUpload)
-
-    return (
-      <div>
-        <Dropzone name={name} style={{}} multiple={false} onDrop={onDrop}>
-            <FlatButton label="Select Wallet File..."
-                                icon={<FontIcon className="fa fa-briefcase" />}/>
-        </Dropzone>
-        {files && <div>Selected: {files[0].name}</div>} 
-        {touched && error && <span className="error">{error}</span>}    
-      </div>
-    )
-}
-
-const renderTextField = ({ input, label, type, meta: { touched, error } }) => (
-  <div>
-    <label>{label}</label>
-    <div>
-      <TextField {...input} type={type} errorText={touched && error} />
-    </div>
-  </div>
-)
 
 const Render = ({account, submitSucceeded, handleSubmit, invalid, pristine, reset, submitting, cancel}) => {
